bugfix: Don't show preparing PC message (#4257)
Previously, we would show "Preparing presentation compiler message", but since it seems to be on a critical path, it sometimes popped up and would stop. Also the current place this is in is not actually downloading anything. Now, I decided after multiple interations to just not show the message. If anything breaks we will show a proper message, but otherwise the message didn't add anything and might cause issues for the users.