fix: MetalsDriver caches only if the target URI is same as the previous run
Now, MetalsDriver skips running compilation if
- the target URI of `run` is the same as the previous target URI
- the content didn't change since the last compilation.

This compilation cache enables Metals to skip compilation and re-use
the typed tree under the situation like developers
sequentially hover on the symbols in the same file without any changes.

Note: we decided to cache only if the target URI only if the same as the previous run
because of `InteractiveDriver.currentCtx` that should return the context that
refers to the last compiled source file.
It would be ideal if we could update currentCtx even when we skip the compilation,
but we struggled to do that. See the discussion
https://github.com/scalameta/metals/pull/4225#discussion_r941138403

To avoid the complexity related to currentCtx,
we decided to cache only when the target URI only if the same as the previous run.