Don't run InteractiveDriver.run if the content didn't "change" (#4225)
* feat: Don't run InteractiveDriver.run if the content didn't change

If the content didn't change since the last `driver.run`, do not compile
because the compilation unit should have the up to date typed tree.
Skip compilation save up much of CPU usage when users are reading
source code (and hover on symbols) without modifying the source code.

What happens if we change in a file affect another file's type
information?

For example, when we have the following source files, if we hover on
`B.foo` in `A.scala`, Metals shows the symbol's type is `Int`.

```scala
// A.scala
def main = println(B.foo)

// B.scala
object B:
  def foo: Int = ???
```

When we modify `B.scala` to `def foo: String = ???`, what happens to
`A.scala`?
You may think that, Metals keep using the old compilation
unit that shows `B.foo` as `Int` instead of `String` because Metals
invalidates the compilation cache based on its content hash.

That's not true, `alreadyCompiled("A.scala", <content hash>)` will be
`false`, and Metals will run `driver.run` on `A.scala` and show `String`
for `B.foo` in `A.scala`.

Why? Because Metals will recreate a new `MetalsDriver` and `lastCompiled`
cache will be cleared up.

- When a file has changed, Metals call [Compilers.restart](https://github.com/scalameta/metals/blob/c6e0f30febeff67e970e463d79862e27e45ada3f/metals/src/main/scala/scala/meta/internal/metals/Compilers.scala#L247).
- The restart method calls `shutdownCurrentCompiler` that [set `_compiler` to `null`.](https://github.com/scalameta/metals/blob/c6e0f30febeff67e970e463d79862e27e45ada3f/mtags/src/main/scala/scala/meta/internal/pc/CompilerAccess.scala#L57-L74).

This is basically the same way to cache compilation as Scala2
See my explanation on how Scala2 + Metals compilation cache works
https://contributors.scala-lang.org/t/how-does-scala2-compiler-cache-invalidate-the-typechecking-result-for-the-given-tree/5858/4?u=tanishiking

* fix: MetalsDriver caches only if the target URI is same as the previous run

Now, MetalsDriver skips running compilation if
- the target URI of `run` is the same as the previous target URI
- the content didn't change since the last compilation.

This compilation cache enables Metals to skip compilation and re-use
the typed tree under the situation like developers
sequentially hover on the symbols in the same file without any changes.

Note: we decided to cache only if the target URI only if the same as the previous run
because of `InteractiveDriver.currentCtx` that should return the context that
refers to the last compiled source file.
It would be ideal if we could update currentCtx even when we skip the compilation,
but we struggled to do that. See the discussion
https://github.com/scalameta/metals/pull/4225#discussion_r941138403

To avoid the complexity related to currentCtx,
we decided to cache only when the target URI only if the same as the previous run.

* Update mtags/src/main/scala-3/scala/meta/internal/pc/MetalsDriver.scala

Co-authored-by: Vadim Chelyshov <qtankle@gmail.com>

* Remove unused imports

Co-authored-by: Vadim Chelyshov <qtankle@gmail.com>