feat: better mill-bsp semanticdb support (#4295)
In the last release of Mill they added better support for producing
SemanticDB for clients like Metals without needing any changes in your
build. However the way it's done is a bit different and unlike the other
build servers meaning that you can't detect it by looking at the
scalacOptions. This changes special-cases Mill for now when a user is
using mill-bsp in a version >= 0.10.6. Mainly this is so users don't see
a big red X mark next to semanticdb in the doctor. There still will be
one for Java support as currently there is no semanticdb being produced
for Java targets in Mill.