bugfix: Rename properly companion object in local scope
Previously, we would not correctly rename companion objects in local contexts due to the fact that it's needed to eactually search for the symbol. Now, we try to correctly identify the companion object in local context based on the current scope.

One situation this might not work when we have the same named case class in an inner scope:

```scala
def method = {
  case class Instance()
  val ins = Instance()
  def inner = {
     case class Instance()
     val ins = Instance()
  }
}
```

however this seems like a rare enough scenario to not overcomplicate things.

We would need to find all the possible same named case classes and establish boundries, which might be quite difficult.

Alternatively, if we manage to make some local renames working with presentation compiler, this will work correctly and might actually be an easier approach.

Fixes https://github.com/scalameta/metals/issues/1539