bugfix: Fix document highlight in for comprehensions
Previously, document highlight in for comprehensions would not highlight both the definition and the usages in yield. This was caused by the fact that these two are really just separate symbols in separate anonymous functions. Now, we try to relate the two owner functions based on their point position, which seems to work correctly.

It's a bit different in case of Scala 3, since there seem to be more trees with the particular symbol generated, but still path will point us to different symbol than the one being used.