fix: Try to download semanticdb even if it might not be supported
https://github.com/scalameta/metals/issues/4449

Previously, requiresSemanticdb will be false if the given
scala version is not yet listed in the `supportedScalaVersion`
which comes from the BuildInfo of Metals.
This means, we have to wait for the next release (or using the snapshot
build) for downloading the semanticdb-plugin for the latest scala
compiler.
(For example, in order to use 2.13.9, we had to wait for the build that
supports scalac 2.13.9).

With this change, Metals try to download the semanticdb-plugin anyway,
and users won't need to wait for the next build for using the latest
scala compiler.