refactor: Move calculating description, kind and label to CompletionValue
Previously, we would match on completion value in multiple places, which makes it hard to modify a behaviour for single type of completion. Now, I moved most logic into specific CompletionValue.

There might be some additional things to refactor, but we can do it later on.

Additionally, this fixes the added type in CaseKeyword completions.