feature: Add support for type definition
This approach uses is a slight modification of definition provider, where we try to additionally find the type symbol of the definition symbol. It seems to work ok and doesn't require too many changes.

In some cases it might not be clear where to go for the type definition, but I think most cases covered by tests make sense. For example the type definition for type is the same as definition.

There is a small bug for named parameters in Scala 2, but that is unrelated to this PR and can be fixed later.