improvement: Try to download mtags for latest supported nightly Scala version
Sometimes, a nigthly version might be released late or use might use an older snapshot of Metals, in which case it will not be possible to use Metals for that nightly. Thanks to this PR, we will find the latest supported nightly version and use it instead. This should work correctly because each nightly can work with classfiles compiled with the same RC.

Unfortunately, this is not possible to test automatically since any version available locally will not be available in snapshots.