refactor(ci): shuffle around build to speed it up
Just curious if this helps because I noticed that the cross tests and
the sbt tests are the two that take the longes. I think we can split out
the scripted tests from the slow sbt tests to speed that one up and I
think we can split out the download dependencies from cross to also
speed that up. While it make sense to have the download with the cross,
the cross still takes around 40 min and then another 10 for the
download. So I'm curious if it's just faster doing this.