bugfix: Fix compliation issues with 3.3.x
Previously, toTextRefinement would return `Closed` type and now it returns `Text`, which is causing the compilation issues.

Now, since the printer prints the types for refinments correctly, we can avoid overriding the method altogether.

The type is printer a bit differently, but it does compile, so let's keep it in line with the compiler.