refactor: ensure we're testing the connection with Mill
When I opened https://github.com/com-lihaoyi/mill/issues/2123 it made me
wonder why we didn't hit on this in our tests, but then I realized we
were actually never testing the initialization with Mill server. This
adds some more tests to ensure we're not only testing the generation of
the BSP config file, but also connection to the build server.