fix: make ivy completion tests more stable
Right now these are failing since a new circe-yaml was released. In the
tests we are trying to completion the 0.14.x series, but that series is
still ongoing, meaning that every new release will break tests. Instead
this change just looks at the 0.13.x series, which is finished. This
will make these tests a bit more stable now and in the future.