refactor: avoid newlines in json doctor output (#4676)
Super small change that I'm currently hitting on in Neovim. When
consuming the json output for the doctor there used to not be any
newlines in the output. Once we added the Java problems we have one
value that has multiple lines. Not all clients will be able to handle
new lines here, so this small change instead just splits on the newline
and gives them back as two items in the array. Again, this only happens
when using the json output.