bugfix: Show completions in parens after newline
Previously, we wouldn't show named argument completions after (\n due to CURSOR being added and not stripped. Now, we properly show those completions.

I additionally had to change the sorting a bit to make NamedArguments show first, which is the same logic as with Scala 2. Together with making those completions symbolic it seems we got almost the same sorting as with Scala 2.