fix: check for the left in diagnosticCode
This was a mistake that I made earlier that I didn't think about.
This can be either a String or an Int and ours are always Strings
so I need to check left for '-1'. The check I had before didn't
make any sense as I was comparing against an either