chore(site): bump docusaurus to 2.2.0
We were on a beta version before so I thought it would be a good idea
to bump up to an actual stable version. Not much changes at all in appearance.
The only notable difference is that you can no lnoger customize the light/dark
toggle without doing a 'swizzle', which is way more work and not worth it
in my opinion. I also needed to add the `appId` for algolia and it looks
like the `customCss` is calculated from a different root since it was erroring
out and not finding it without changing what we had