improvement: Allow compiler rename in lambdas
Previously, rename would not work using the presentation compiler inside lambdas, which should not require semanticdb for rename. Now, we also check if one of the owners of the renamed symbol is an anonymous function, since that symbol will be local.

For Scala 3 this seems to work correctly already since `ow.is(Method)` is true for lambdas.