refactor: improve reporting on Semanticdb from Mill for Java Modules
This is a follow-up to the work that I've done in
https://github.com/scalameta/metals/pull/4295. At that time I only
applied the changes to Scala Targets since Semanticdb was only produced
there when using Mill as your build server. Now that https://github.com/com-lihaoyi/mill/pull/2227
is merged Java Targets and mixed targets will now also produce
semanticdb. We now add the extra check for Mill and its version when we
check for semanticdb java support. This will ensure that users won't get
the error in the Doctor about lacking support.