refactor: Split DebugAdapterStart command into separate commands
Previously, we would have one command and we would parse the arguments to figure out what we need to do. This meant we couldn't use the ParametrizedCommand etc. and this one command had multiple behaviours. Now, I split all possible arguments into separate commands, which share the id, but thanks to using unapply of ParametrizedCommand we are able always end up at the correct case.

One issue is that we need to keep the in the same order, but this was exactly the same previously. As an added bonus I had to document all the possible commands we have.