improvement: Fallback to symbol search as the last possibility for definition (#4861)
improvement: Fallback to symbol search as the last possibility for defition

Previously, if the code never compiled, we would miss semanticdb and we often wouldn't be able to navigate to a symbol from outside the current file. Now, instead we will try to use the symbol search to find the symbol matching the current name.

This solution might cause some false positives with the same names, but I think overall this might help users a lot and avoid having broken metals in case of not compiling workspace.