improvement: Allow manual cancel of starting debug session
Previously, users would need to wait until the session times out, which would mean that it might be too long for some users and too little for some other. Now, we offer to cancel the session within a slow task.

This gives the control to the user instead of arbitrary timeout.

Fixes https://github.com/scalameta/metals/issues/4864