improvement: Cancel all in BspSession
Previously, we would only invoke shutdown on the build server session and that would not stop the ongoing requests. Now, we also cancel all the request.

I saw a couple of situations where Metals server would stay up unable to shutdown. Not sure if this was the reason, but it should not cause any issues.