chore: Choose object when using symbol search as fallback
Previously, we would suggest both object and class even if it clearly could not be a class. Now, we add some heuristics to minimize false positives.

We do it twofold:
- object will be always identified if we are in a select statement such as `Main.run()` or `val a: Main.Bar`
- objects will not be accepted if we are at a type position

Also, we will now correct search the workspace based on the current file, so if something is not accessible from current file it will not be found.