fix(ci): don't run scripted tests twice
It looks like we're currently running out scripted tests twice when we
don't need to. We're running it in the sbt integration job and also in
the sbt-metals/scripted job. This remotes the scripted stuff from the
sbt integration job and just relies on it running in the scripted one.