bugfix: navigation for apply function in a companion object
Previously: ScalaMtags indexer would not take notice of the automatically created apply function that is produced in a companion object for a case class. This lead to the disambiguator counter to be off by one and goto pointing to the wrong reference.
Now: we make sure that classes are indexed before objects and record found case class to fix the disambiguator counter for apply.