improvement: Add the possibility to only support running the code
Previously, to run the code clients would need to support DAP, which is quite heavywight and not needed at all times. Now, we add `runProvider` that can be used to indicate that the editor only wants to support `run` lenses. The editor can later use the `shellCommand` field to run the code.