improvement: Merge input and input3 projects used for tests
Previously, we would use input with Scala 2 compatible files in Scala 2 specific test and input3 with different files for Scala 3 tests.

Now, we merged all the sources with the incompatible ones in scala-2 or scala-3 folders. We still have two projects, but we will test more files from both Scala 3 and Scala 2 perspective.