bugfix: Don't collect symbols from Inline expansion
Previously, Metals would have a hanging thread when for example in github.com:Atry/scalajs-all-in-one-template. Now, it doesn't end up in an infinite loop.

I haven't been able to find a proper test for it aside from the very complex example with macros, but we shouldn't need to traverse expansion as it's the tree that is not in the original code.