docs: Switch from algolia to easyops-cn/docusaurus-search-local
We've been working on updating the algolia index for scalameta sites.
However, there's been a lot of trouble keeping the algolia keys up to date,
and the search function sometimes goes down.
see:
https://github.com/scalacenter/scalafix/issues/1514
https://github.com/scalameta/metals/pull/3604
https://github.com/scalameta/scalameta/pull/2569

This commit switches the search function from algolia to one of the
local search indexes
https://github.com/easyops-cn/docusaurus-search-local
chosen from
https://docusaurus.io/community/resources#search

I picked easyops-cn/docusaurus-search-local because

- it seems to be well maintained
- it works without server setup (unlike typesense)
- looks prettier than cmfcmf/docusaurus-search-local (original one)

but, other suggestions are welcome :)