bugfix: Fix semantic highlight in sbt files
Previously, semantic highlighting would be broken inside Metal's build.sbt file. This was caused by the fact that start/end is validated to always be `start <= end`, so we need to change it together instead one after another. Another issue was if we haven't find the literal name inside the source code we would get another exception.

Now, both cases are handled and this was tested manually for metals build file. This wasn't breaking in smaller test cases.

Also, when no semantic tokens are returned, we return an empty list not to interfere with the default highlighting.