bugfix: Fix go to definition for TypeTest
In some case TypeTest will be used in pattern match and it actually changes the case to a new unapply. Previously, we wouldn't be able to figure out the initial type.

Now, we special case TypeTest and find the exact typeSymbol we need.

Connected to https://github.com/scalameta/metals/issues/5102