bugfix: Show document highlight with cursor on primary cosntructor
Previously, when using document highlight on the type parameter in the primary constructor, we would only get that specific parameter highlighted. This is because those were treated as different symbols with one having constructor parent and the other trait param.

Now, we properly match them as the same symbol from the user perspective.

This seems to be handled the other way already.