refactor: account for new diagnosticData shape
Previously in scala-cli the diagnostic `data` for actionable diagnostics
were `textEdit`s, whereas now they are nested under `edits` and are
`workspaceEdit`s. This change will ensure they still work for the old
format and also for the new.

This relates to the changes in https://github.com/VirtusLab/scala-cli/pull/1969.