refactor: use -release isntead of -target
```
The -release option specifies the target version, such as “8” or “18”.

Like the option for javac, it allows building against an earlier version of the JDK. It will compile against the API for that version and also output class files for that version.

The deprecated option -target does not compile against the desired API, but only specifies a target class file format.
```

This allows us to avoid issues with breaking compilation on older JVMs.

I also merged the other options and remove `-Yrangepos` for Scala 3