Keep exception cause in some mtags exceptions
Passing the cause exception to the Exception contructor makes it easier
for users to know what's going on: the JVM or the printStackTrace method
will print the current exception and its cause (and the cause's cause, etc.).
That way, we know in more detail what the underlying cause of the exception
is.