bugfix: Fix issues when shortened type is used in type bounds
Previously, if we would try to show shorter type without packages in type bounds we would get an exception since type bounds require TermType.

Now, I switched TermType to be used for Pretty type and changed to default to long name in case of errors. I haven't seen issues related to TypeType, so we should be fine there, but we will get a nice error reprot if this happens.

Fixes https://github.com/scalameta/metals/issues/5225