refactor: a bit more details in logs about whether the java path exists (#5401)
refactor: make info about java binary debug

Super tiny change. I just saw this in the logs today and wasn't 100%
what the boolean value printed out was meaning until I looked at the
code. This just adds a tiny change to the logs to make it explicit that
the boolean being printed out refers to whether or not the path exists.

It also changes it from info to debug.