improvement: Automatically add colon to get better completions for Java
Previously, if there was anything following the place you were trying to get a completion in, you would not get any suggestions since the following names would be take into account. Now, we add `;` after anything you are trying to complete if the next symbol is a whitespace to separate the current completion from the following statements.

I also noticed that we don't set proper snippet format for the items, so I added that.