bugfix: Fix issues with select in debugger completions
Previously, we set the wrong start of the expression since we actually need to calculate that when insertText is used. Now, we calculate it based on the current identifier.

We should use TextEdit in the future, which should work better. I also realized that start actually needs to take into account if the client starts at 1 or 0.