improvement: Base Metals view on indexing information
Previously, we would read the classpath to create the metals view, which worked well for Scala 2, but needed a reimplementation for Scala 3. So now, instead of doing this reimplementation, which would probably need to use tasty and might be quite complex, I decided to base it on the same approach we use for inexing.

This makes the feature more reliable since, we now used the same well tested functionaties as things like go to definition.

Unfortunately, we might index the particular files the second time, because the basic indexing caches are reversed and finding all symbols in a file from those maps was much slower than just doing it the second time. We later cache that information the same way we did in the previous tree view implementation. So we won't use more memory and running the benchmarks show that the new approach is similarily fast:

```
[info] Benchmark                  Mode  Cnt    Score    Error  Units
[info] ClasspathSymbolsBench.run    ss   10  542.090 ± 61.953  ms/op
```

as compared to the previous:
```
[info] Benchmark                  Mode  Cnt    Score    Error  Units
[info] ClasspathSymbolsBench.run    ss   10  477.283 ± 93.933  ms/op
```

Also it seems we index much more information, since mtags index all symbols, which would also explain the slight difference in timings as I think the difference was 1300 symbols previously and 17000 after this change (number of symbols detected in the benchmarks)

Fixes https://github.com/scalameta/metals/issues/2859