bugfix: Don't enable * kind project syntax by default
Previously, the default Scala3 dialect would assume kind projector syntax, but that is only true if the proper setting is on.

Now, we only turn it on if defined in scalac options to be on a safe side. There is no current features that depend on it, but it's probably better to have the right trees.

Fixes https://github.com/scalameta/metals/issues/5573