Bump swoval to 2.1.12
I discovered a small regression that caused sbt's tests to fail in
2.1.11 (https://github.com/sbt/io/pull/352). The 2.1.12 release
fixes the regression (https://github.com/swoval/swoval/pull/173) while
still making the switch from the jni based directory lister to the
nio based directory lister on windows.  It is possible that 2.1.12 is
slower than 2.1.11 on Windows because it restores some additional
directory scanning during registration. My intuition is that it should
not go back up to 14 minutes for the reporting user's use case but it
might not be as quick as 10 seconds.