improvement: Automatically calculate last supported versions for semanticdb
Previously, we would fail if our version of semanticdb was not supported for a sepcific scala version. Now, we calculate last supported version for each scala version when compiling.

Potentially this means we could get a new semanticdb version automatically, but that should not be an issue as semanticdb is well tested in scalameta itself