bugfix: Don't read file for Interactivedb if it doesn't exist
Previously, we would assume that any file that we want to calculate InteractiveSemanticdb for exists on the filesystem, which is not true. For example renamed files if renamed from terminal will stick in the editor and cause exceptions to be throw. Now, we also try to read it from buffers and from filesystem as a fallback if it exists on the filesystem.