refactor: differentiate between files and directories in sources
So I noticed that when using the "Show build target info" feature while
using scala-cli it would show individual files under the "Source
Directories". I know it's a bit of a nit, but I thought it might be
nicer to rename this to "Sources" and then do a better job at indicating
which are files, which are directories, and which ones don't exist. If
we agree this is better I can fix the related tests as well.