improvement: Send error back to client to show
Previously, it would be very difficult to show a proper error message to the user. Especially in VS Code a generic message was show, which was very confusing for users.

Now, we send the error message in the response so that the client can handle it as they want.

I wonder if we should keep displaying the message, but not sure if it's not better to just change the clients here to show the error.