fix: use client commands inside of BSP status
These commands that are attached to status' are meant for the client to
execute and send to the server. I noticed this because in nvim-metals
I've gotten multiple reports of these commands not working when they
arrive since normally the client commands are prefaced with `metals-`
whereas the server ones aren't. I'm not sure how this was working in VS
Code or if it was, but I believe this change is correct.