feat: cancel worksheet evaluations when canceling compilation (#6013)
This is sort of a hack, but this is meant to be for clients that don't
have slowTask abilities, which means they can't cancel long-running
tasks. In the situation where you're in a worksheet and have something
that you want to cancel, this allows you to trigger a cancel
compilation, which will do that and also cancel the ongoing evaluation
if the user is in a worksheet.