bugfix: Fix displaying doctor information using bazel
Previously, when Bazel targets included `/` which is almost always, we would not be able to find the build target to display in the doctor. Now, I switched / to : for the link.

It's incredibly tricky to get / working correctly in URIs, so it's much easier to just replace it.

I also added a couple small fixes that help out with local testing of Bazel BSP