chore: Change to slf4j2 to avoid too much logging in flyway
Previously, we would get all the messages up to debug level from flyway, which added a lot of overhead. It seems that this is due to the fact that flyway has issues detecting slf4j loggers and especially scribe, which implements it. Changing to slf4j2 helped here and we get much less data.