improvement: Use dependencyModule for resolving sources
Previously, we would assume that the source jar is in the same directory as the normal jar. Now we use dependencyModules for that and only fallback to the previous logic if we can't find the jar in dependency modules.

This extracts part of the lazy classpath PR

Fixes https://github.com/scalameta/metals/issues/6182