improvement: Simplify resolving presentation compiler
Previously, we would have a lot of fallbacks that would be used if mtags was not avaiable. Now, each version of the compiler has a stable version of the compiler being released so there is no longer a reason to handle it via fallback. Instead we want to be explicit about what we are resolving.

It seems that previous way was flaky and while I can't reproduce the issues people had, I think simplifying it would help out at lest to get the exact failure of what is going on.