Fixed issue where oldPath wasn't using OS specific spelling for Path/PATH
For Windows users oldPath would always return as an empty string. This is because we're using "PATH" instead of "Path".
Being's that would always return as an empty string, the newPath would always be the javaBin string.
When storing in the map, we used the OS specific Path/PATH but with the wrong information.