improvement: Recreate classloader if scalafix classloading fails
I wasn't able to figure out why this might be happening for people, but usually this happened if scalafix was invoked at a wrong time, running it later would work.

I also added reporting and a potential fix for Scalafix cache, since we could have a race updating caches and with getOrElseUpdate we are at least sure the same exact result will be returned.