feature: Add support for Scala 3's Best Effort compilation (#5219)
* feat: Add Best Effort compilation support for Scala 3

This PR is meant to add support for Best Effort compilation in metals,
with a few additional considerations:
* Some unnecessary options are replaced from bloop and set for PC
* Previously, any project could be run from metals if semanticdb was
able to be generated. This caused a somewhat weird UX for best effort.
Now the build directory is explicitly checked for the existence of non
META-INF files.
* Previously, in metals doctor, compilation would be set as correct if
a compilation did not produce any errors. However, with best effort
compilation, the compiler may not find any additional errors, beyond
the ones from its dependencies (meaning that a project that cannot
actually be compiled, could show up as on that was compiling). A note
about that was added in the doctor.

* Setup previous outline tests so they are alos run for best effort

* Address review comments

Now we check diagnostics of all transitive dependencies for errors,
instead of checking the existence of non meta-inf files.

Also the keep-after-error test now does not check for remaining code
lens after saving.