bugfix: Only detect 100% sure test names
Previously, we would treat any string literal inside the apply of a test function as a name of such function, but each test name can actually be any kind of expression, so we would not detect the correc tname always.

Now, we only detect names of tests if they are of an expected format, either simple string, select or a tag.

Fixes https://github.com/scalameta/metals/issues/6617