improvement: Improve heuritics used for fallback symbol search
Previously, we would just use name when nothing else was working, which could lead to a lot of false positives.

Now, we take into account imports, current package and fully qualified names to find the correct definition.

One case this might not work is with unimports, but that should not be a big deal as both will be found only if nothing was ever compiled.