improvement: Detect custom mains
Previously, if any plugin generate a separate main entry we would not show it in the code lenses. Now, we show it at the start of the file, since without an occurrence we can't say where the code lens should end up.

This might be improved in the future if we add those symbols to synthetic section, but that is much more work.