refactor: Use codeAction interface for most code actions
I think we were still not perfect with the interfaces, some changes I did:
- I changed the default value to empty, which now makes it easy to see which methods migrated to the new interface method.
- removed maybeCodeActionId from code actions that we know can be handled otherwise
- added a CodeActionId class to gather all known classes once we agree on adding soemthing, can still be implemented separately

This added some duplication, since for older Scala version we need to use the older methods, but anything new should now be quite easy to handle.

So not so useful for older things, but still it's good to migrate to new codeAction interface.