bugfix: Add name field to distinguish different action data
Also refactor another code action to detect possible conflicts

I was hoping to have it cleaner, but using a stable id between codeAction request and resolve one is safest. I also check if everything is not null, so we should be safe even if someone forgets to set the name properly.