improvement: Make including detail in completion label configurable
Some LSP clients (such as Emacs using Eglot and Corfu) use the label field to insert a completion if
completionItem/resolve has not returned by the time the completion is selected. Including the contents of the detail
property in the label causes it to be included in the file itself in these clients. To remedy this, we define a
isDetailIncludedInLabel option under compilerOptions to make this behaviour toggleable. We also disable this behaviour
by default in Emacs, since this is where the undesirable behaviour was located.

Fixes #6849.