bugfix: Remove pausing and window state monitoring
I finally managed to reproduce an issue where not compilation would be done, turns out the batched functions went into a wrong state and never unpaused. In that case any request would be silently ignored.

Since, this looks tricky to make work and the advantage is minimal I think it's best to remove it altogether. It could be useful if we had two heavy compilations one after another, but that very rarely happens, since the next compilation is usually noop.