bugfix: Fix test incompatibility
It seems pc definition will not find Seq, but will try to find apply on `Seq("")`. With the current changes when cursor is on ( we will still try to find the definition and will end up on the apply. In the previous solution we would return empty and semanticdb definition will return Seq located in scala/package