bugfix: Fix when types are coming from package objects
It seems there is some ordering issues when the current package and the one from the package object are the same.

The types from that package object are never put into scope, since it seems it's usually done lazily, but for the current package it's ignored.

I haven't managed to get to the total bottom of it, but I suspect that it has someting to do with the multiple lazy analyzers in the inheritance of Global.

Fixes https://github.com/scalameta/metals/issues/4495