bugfix: don't prefix scope completions from supertype
In the added test case, the previous behaviour would've been to complete
to `Foo.overloaded` instead of just `overloaded`. This was because the
effective owner of `overloaded` was `Foo` instead of `Bar`. Since those
were not the same symbol, the logic for determining the prefix would
incorrectly add a `Foo` prefix to the completion.

By augumenting the equality check to be a subtyping check, we ensure
that we don't add prefixes to completions coming from supertype scopes.