improvement: Only show symbol import if static method exists
For now it's a small improvement, I tried some other cases, but for example if we have rhs that we want to check the type for it's currently impossible when the lhs type is not imported.
Inserting the import and rechecking would be too costly in the auto imports context unfortunately.