bugfix: Make sure to choose the best import option
When importing symbols in a range it might turn out we are able to get a better suggestions in a certain spot in the file.

For example if we know that one object has a specific method while the other doesn't.

This means that this import is better for the entire file since it's more concrete.

Previously, we would still show all options, but now search for the shortest action choice for each missing symbol.